Rename unittest.py to unittestplugin.py
Renaming to avoid confusion with unittest in standard Python library.
