+++ spyder_unittest/__init__.py
@@ -8,7 +8,7 @@
 """Spyder unitest plugin."""
 
 # Local imports
-from .unittest import UnitTestPlugin as PLUGIN_CLASS
+from .unittestplugin import UnitTestPlugin as PLUGIN_CLASS
 
 VERSION_INFO = (0, 1, 0, 'dev0')
 __version__ = '.'.join(map(str, VERSION_INFO))
